feat: Remove redundant code and add rpc-starter(no work)
