+++ spring-cloud-alibaba-dependencies/pom.xml
@@ -34,14 +34,16 @@
         <shedlock.version>4.23.0</shedlock.version>
         <schedulerx.worker.version>1.11.4</schedulerx.worker.version>
 
+        <!-- dubbo-->
+        <dubbo.version>3.3.0-beta.4</dubbo.version>
+
         <!-- Maven Plugin Versions -->
         <maven-source-plugin.version>3.2.1</maven-source-plugin.version>
         <maven-javadoc-plugin.version>3.1.1</maven-javadoc-plugin.version>
         <maven-gpg-plugin.version>3.0.1</maven-gpg-plugin.version>
         <flatten-maven-plugin.version>1.2.7</flatten-maven-plugin.version>
 
-        <!-- dubbo-->
-        <dubbo.version>3.3.0-beta.4</dubbo.version>
+
     </properties>
 
     <dependencyManagement>
@@ -208,7 +210,12 @@
                 <artifactId>schedulerx2-worker</artifactId>
                 <version>${schedulerx.worker.version}</version>
             </dependency>
-
+            <!--  dubbo Dependencies  -->
+            <dependency>
+                <groupId>org.apache.dubbo</groupId>
+                <artifactId>dubbo</artifactId>
+                <version>${dubbo.version}</version>
+            </dependency>
             <!-- Own dependencies  -->
             <dependency>
                 <groupId>com.alibaba.cloud</groupId>
@@ -287,6 +294,11 @@
                 <artifactId>spring-cloud-starter-alibaba-schedulerx</artifactId>
                 <version>${revision}</version>
             </dependency>
+            <dependency>
+                <groupId>com.alibaba.cloud</groupId>
+                <artifactId>spring-cloud-starter-alibaba-rpc</artifactId>
+                <version>${revision}</version>
+            </dependency>
 
             <dependency>
                 <groupId>com.alibaba.spring</groupId>
@@ -300,13 +312,6 @@
                 <artifactId>spring-ai-test</artifactId>
                 <version>${spring.ai.version}</version>
                 <scope>test</scope>
-            </dependency>
-
-            <!--  dubbo Dependencies  -->
-            <dependency>
-                <groupId>org.apache.dubbo</groupId>
-                <artifactId>dubbo-spring-boot-starter</artifactId>
-                <version>${dubbo.version}</version>
             </dependency>
         </dependencies>
     </dependencyManagement>
