+++ spring-cloud-alibaba-starters/spring-cloud-starter-alibaba-rpc/src/main/java/com/alibaba/cloud/rpc/constant/RpcConstant.java
@@ -1,0 +1,10 @@
+package com.alibaba.cloud.rpc.constant;
+
+/**
+ * @author :Lictory
+ * @date : 2024/09/20
+ */
+public class RpcConstant {
+
+    public static final String SPRING_RPC_NETTY_PORT = "spring.rpc.netty.port";
+}
