+++ spring-cloud-alibaba-starters/spring-cloud-starter-alibaba-rpc/src/main/java/com/alibaba/cloud/rpc/server/RpcDispatcherServlet.java
@@ -1,0 +1,17 @@
+package com.alibaba.cloud.rpc.server;
+
+import com.alibaba.cloud.rpc.utils.SpringUtil;
+import org.springframework.web.servlet.DispatcherServlet;
+
+/**
+ * @author :Lictory
+ * @date : 2024/08/31
+ */
+
+public class RpcDispatcherServlet extends DispatcherServlet {
+    public RpcDispatcherServlet() {
+        super();
+        // 初始化mapping
+        onRefresh(SpringUtil.applicationContext);
+    }
+}
