--- spring-cloud-alibaba-starters/spring-cloud-starter-rpc/pom.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>com.alibaba.cloud</groupId>
-        <artifactId>spring-cloud-alibaba</artifactId>
-        <version>2023.0.1.2</version>
-        <relativePath>../../pom.xml</relativePath>
-    </parent>
-
-    <groupId>org.example</groupId>
-    <artifactId>spring-cloud-starter-rpc</artifactId>
-
-    <properties>
-        <maven.compiler.source>19</maven.compiler.source>
-        <maven.compiler.target>19</maven.compiler.target>
-        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    </properties>
-
-</project>
