--- spring-cloud-alibaba-tests/rpc-tests/spring-cloud-alibaba-rpc-core/src/main/java/com/alibaba/rpc/common/RpcConstance.java
@@ -1,9 +1,0 @@
-package com.alibaba.rpc.common;
-
-/**
- * @author :Lictory
- * @date : 2024/08/19
- */
-public interface RpcConstance {
-    String LOCAL_HOST_HEADER = "http://localhost:";
-}
