--- spring-cloud-alibaba-tests/rpc-tests/spring-cloud-alibaba-rpc-core/src/main/resources/META-INF/spring.factories
@@ -1,3 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-  com.alibaba.dubbo.consumer.config.RpcFeignConfiguration,\
-  com.alibaba.dubbo.consumer.config.RpcRegistryConfiguration,\
