--- spring-cloud-alibaba-tests/rpc-tests/spring-cloud-alibaba-rpc-core/src/main/resources/application.yml
@@ -1,1 +1,0 @@
-
