--- spring-cloud-alibaba-tests/rpc-tests/spring-cloud-alibaba-rpc-provider/src/main/java/com/alibaba/dubbo/provider/handler/RpcNettyServerHandler.java
@@ -1,24 +1,0 @@
-package com.alibaba.dubbo.provider.handler;
-
-/**
- * @author :Lictory
- * @date : 2024/08/15
- */
-import io.netty.channel.ChannelHandlerContext;
-import io.netty.channel.ChannelInboundHandlerAdapter;
-
-public class RpcNettyServerHandler extends ChannelInboundHandlerAdapter {
-
-    @Override
-    public void channelRead(ChannelHandlerContext ctx, Object msg) {
-        System.out.println("Server received: " + msg);
-        ctx.writeAndFlush("Message from server: " + msg);
-    }
-
-    @Override
-    public void exceptionCaught(ChannelHandlerContext ctx, Throwable cause) {
-        cause.printStackTrace();
-        ctx.close();
-    }
-}
-
