+++ spring-cloud-alibaba-tests/rpc-tests/spring-cloud-alibaba-rpc-provider/src/main/resources/application.yml
@@ -1,3 +1,4 @@
+
 spring:
   main:
     allow-bean-definition-overriding: true
@@ -7,11 +8,15 @@ spring:
     nacos:
       discovery:
         server-addr: 10.21.32.154:8848
-  port: 8080
-
+  rpc:
+    enable: true
+    netty:
+      host: 127.0.0.1
+      port: 20880
 management:
   endpoints:
     web:
       exposure:
         include: "*"
-
+server:
+  port: 8080
